Split 'AnswerGrid' into multiple files
