+++ newdle/client/src/components/answer/AnswerGrid.js
@@ -1,26 +1,15 @@
-import React, {useState} from 'react';
+import React from 'react';
 import {useDispatch, useSelector} from 'react-redux';
-import {Trans} from '@lingui/macro';
 import _ from 'lodash';
 import moment from 'moment';
 import PropTypes from 'prop-types';
-import {Icon, Table, Radio, Popup} from 'semantic-ui-react';
+import {Icon, Table} from 'semantic-ui-react';
 import {setAnswer} from '../../actions';
 import {getAnswers} from '../../answerSelectors';
-import {
-  getNewdleDuration,
-  getNewdleParticipants,
-  getNewdleTimezone,
-  getNewdleTimeslots,
-} from '../../selectors';
-import {serializeDate, toMoment} from '../../util/date';
-import AvailabilityRing from '../AvailabilityRing';
-import styles from './AnswerGrid.module.scss';
-
-function formatMeetingTime(startTime, duration) {
-  const endTime = moment(startTime).add(duration, 'm');
-  return `${serializeDate(startTime, 'HH:mm')} - ${serializeDate(endTime, 'HH:mm')}`;
-}
+import {getNewdleParticipants, getNewdleTimeslots} from '../../selectors';
+import {toMoment} from '../../util/date';
+import {NameCell, TableHeader, TableFooter} from '../GridCommon';
+import styles from '../GridCommon.module.scss';
 
 /**
  * Computes the width and offset of the red strips used in the answer grid view,
@@ -64,159 +53,6 @@ function getBusyTimesPositions(timeslot, busyTimes, duration) {
 
   return styles;
 }
-
-function FooterCell({
-  participants,
-  timeslot,
-  interactive,
-  active,
-  hovered,
-  onMouseEnter,
-  onMouseLeave,
-}) {
-  const availableCount = participants.filter(({answers}) => answers[timeslot] === 'available')
-    .length;
-  const unavailableCount = participants.filter(({answers}) => answers[timeslot] === 'unavailable')
-    .length;
-  const ifneedbeCount = participants.filter(({answers}) => answers[timeslot] === 'ifneedbe').length;
-
-  const className = active ? styles.active : hovered && interactive ? styles.hover : null;
-
-  return (
-    <Table.HeaderCell
-      textAlign="center"
-      className={className}
-      onMouseEnter={onMouseEnter}
-      onMouseLeave={onMouseLeave}
-    >
-      <AvailabilityRing
-        available={availableCount}
-        unavailable={unavailableCount}
-        ifNeeded={ifneedbeCount}
-        totalParticipants={participants.length}
-        radius={40}
-        strokeWidth={7}
-      />
-    </Table.HeaderCell>
-  );
-}
-
-FooterCell.propTypes = {
-  participants: PropTypes.arrayOf(
-    PropTypes.shape({
-      answers: PropTypes.objectOf(PropTypes.oneOf(['unavailable', 'available', 'ifneedbe'])),
-    })
-  ).isRequired,
-  timeslot: PropTypes.string,
-  interactive: PropTypes.bool.isRequired,
-  active: PropTypes.bool.isRequired,
-  hovered: PropTypes.bool.isRequired,
-  onMouseEnter: PropTypes.func,
-  onMouseLeave: PropTypes.func,
-};
-
-function DateCell({
-  timeslot,
-  setFinalDate,
-  isCreator,
-  interactive,
-  active,
-  hovered,
-  onMouseEnter,
-  onMouseLeave,
-}) {
-  const newdleTimezone = useSelector(getNewdleTimezone);
-  const duration = useSelector(getNewdleDuration);
-
-  const startTime = toMoment(timeslot, 'YYYY-MM-DDTHH:mm');
-
-  let className = styles.header;
-  if (active) {
-    className += ` ${styles.active}`;
-  } else if (hovered && interactive) {
-    className += ` ${styles.hover}`;
-  }
-
-  if (isCreator && interactive) {
-    className += ` ${styles.pointer}`;
-  }
-
-  return (
-    <Table.HeaderCell
-      textAlign="center"
-      className={className}
-      onClick={() => interactive && isCreator && setFinalDate(timeslot)}
-      onMouseEnter={onMouseEnter}
-      onMouseLeave={onMouseLeave}
-    >
-      <div>
-        <div className={styles['date']}>{startTime.format('D MMM')}</div>
-        <div className={styles['time']}>{formatMeetingTime(startTime, duration)}</div>
-        <div className={styles['timezone']}>{newdleTimezone}</div>
-      </div>
-      {interactive && isCreator && <Radio name="slot-id" value={timeslot} checked={active} />}
-    </Table.HeaderCell>
-  );
-}
-
-DateCell.propTypes = {
-  timeslot: PropTypes.string,
-  setFinalDate: PropTypes.func,
-  isCreator: PropTypes.bool.isRequired,
-  interactive: PropTypes.bool.isRequired,
-  active: PropTypes.bool.isRequired,
-  hovered: PropTypes.bool.isRequired,
-  onMouseEnter: PropTypes.func,
-  onMouseLeave: PropTypes.func,
-};
-
-function NameCell({participant: {avatar_url, comment, name}, highlighted = false}) {
-  let avatarURL;
-  if (avatar_url) {
-    avatarURL = new URL(avatar_url, window.location.origin);
-    avatarURL.searchParams.set('size', 30);
-  }
-
-  const trigger = (
-    <Table.Cell className={`${styles['avatar-cell']} ${highlighted ? styles.highlighted : ''}`}>
-      {comment && (
-        <Icon
-          corner="top left"
-          flipped="horizontally"
-          name="comment outline"
-          className={`${styles['middle-aligned']} ${styles['comment-icon']}`}
-        />
-      )}
-      {avatarURL ? (
-        <div className={`${styles['middle-aligned']} ${styles.avatar}`}>
-          <img className="user-avatar" src={avatarURL} alt="" />
-        </div>
-      ) : (
-        <Icon
-          size="big"
-          name="user circle"
-          className={`${styles['middle-aligned']} ${styles.avatar}`}
-        />
-      )}
-      <span className={`${styles['middle-aligned']} ${styles.name}`}>{name}</span>
-    </Table.Cell>
-  );
-
-  return comment ? (
-    <Popup wide position="top center" mouseEnterDelay={100} trigger={trigger} content={comment} />
-  ) : (
-    trigger
-  );
-}
-
-NameCell.propTypes = {
-  participant: PropTypes.shape({
-    avatar_url: PropTypes.string,
-    comment: PropTypes.string,
-    name: PropTypes.string.isRequired,
-  }),
-  highlighted: PropTypes.bool,
-};
 
 function AnswerCell({
   participant,
@@ -295,68 +131,6 @@ AnswerCell.propTypes = {
   unknown: PropTypes.bool.isRequired,
 };
 
-function SummaryCell({
-  participant,
-  timeslot,
-  interactive,
-  hovered,
-  active,
-  onMouseEnter,
-  onMouseLeave,
-  hasAnswered,
-}) {
-  const status = participant.answers[timeslot];
-  const positive = status === 'available';
-  const negative = status === 'unavailable';
-
-  const statusColors = {available: 'green', ifneedbe: 'yellow', unavailable: 'red'};
-
-  const icon = status ? (
-    <Icon
-      name={status !== 'unavailable' ? 'checkmark' : 'close'}
-      color={statusColors[status]}
-      size="large"
-    />
-  ) : hasAnswered ? (
-    <Icon name="question" color="grey" size="large" />
-  ) : null;
-
-  const className = active
-    ? styles.active
-    : hovered && interactive
-    ? styles.hover
-    : !interactive
-    ? styles.finalized
-    : null;
-
-  return (
-    <Table.Cell
-      positive={positive}
-      negative={negative}
-      key={timeslot}
-      textAlign="center"
-      className={className}
-      onMouseEnter={onMouseEnter}
-      onMouseLeave={onMouseLeave}
-    >
-      {icon}
-    </Table.Cell>
-  );
-}
-
-SummaryCell.propTypes = {
-  participant: PropTypes.shape({
-    answers: PropTypes.objectOf(PropTypes.oneOf(['unavailable', 'available', 'ifneedbe'])),
-  }).isRequired,
-  timeslot: PropTypes.string,
-  interactive: PropTypes.bool.isRequired,
-  active: PropTypes.bool.isRequired,
-  hovered: PropTypes.bool.isRequired,
-  onMouseEnter: PropTypes.func,
-  onMouseLeave: PropTypes.func,
-  hasAnswered: PropTypes.bool.isRequired,
-};
-
 function AnswerRow({
   participant,
   timeslots,
@@ -400,136 +174,7 @@ AnswerRow.propTypes = {
   selectable: PropTypes.bool.isRequired,
 };
 
-function SummaryRow({
-  participant,
-  timeslots,
-  interactive,
-  finalDate,
-  hoveredColumn,
-  setHoveredColumn,
-}) {
-  const hasAnswered = Object.keys(participant.answers || []).length !== 0;
-
-  return (
-    <Table.Row textAlign="center">
-      <NameCell participant={participant} />
-      {timeslots.map(timeslot => (
-        <SummaryCell
-          key={timeslot}
-          participant={participant}
-          timeslot={timeslot}
-          interactive={interactive}
-          hovered={interactive && hoveredColumn === timeslot}
-          active={interactive && finalDate === timeslot}
-          onMouseEnter={interactive ? () => setHoveredColumn(timeslot) : null}
-          onMouseLeave={interactive ? () => setHoveredColumn(null) : null}
-          hasAnswered={hasAnswered}
-        />
-      ))}
-    </Table.Row>
-  );
-}
-
-SummaryRow.propTypes = {
-  participant: PropTypes.shape({
-    answers: PropTypes.objectOf(PropTypes.oneOf(['unavailable', 'available', 'ifneedbe'])),
-    name: PropTypes.string.isRequired,
-    email: PropTypes.string,
-    comment: PropTypes.string,
-  }).isRequired,
-  timeslots: PropTypes.array.isRequired,
-  interactive: PropTypes.bool.isRequired,
-  finalDate: PropTypes.string,
-  hoveredColumn: PropTypes.string,
-  setHoveredColumn: PropTypes.func.isRequired,
-};
-
-function TableHeader({
-  timeslots,
-  interactive,
-  finalDate,
-  setFinalDate,
-  isCreator,
-  hoveredColumn,
-  setHoveredColumn,
-}) {
-  return (
-    <Table.Header>
-      <Table.Row>
-        <Table.HeaderCell textAlign="center">
-          <Trans>Participants</Trans>
-        </Table.HeaderCell>
-        {timeslots.map(timeslot => (
-          <DateCell
-            key={timeslot}
-            timeslot={timeslot}
-            interactive={interactive}
-            setFinalDate={setFinalDate}
-            isCreator={isCreator}
-            hovered={interactive && hoveredColumn === timeslot}
-            active={interactive && finalDate === timeslot}
-            onMouseEnter={interactive ? () => setHoveredColumn(timeslot) : null}
-            onMouseLeave={interactive ? () => setHoveredColumn(null) : null}
-          />
-        ))}
-      </Table.Row>
-    </Table.Header>
-  );
-}
-
-TableHeader.propTypes = {
-  timeslots: PropTypes.array.isRequired,
-  interactive: PropTypes.bool.isRequired,
-  finalDate: PropTypes.string,
-  setFinalDate: PropTypes.func,
-  isCreator: PropTypes.bool.isRequired,
-  hoveredColumn: PropTypes.string,
-  setHoveredColumn: PropTypes.func,
-};
-
-function TableFooter({
-  participants,
-  timeslots,
-  interactive,
-  finalDate,
-  hoveredColumn,
-  setHoveredColumn,
-}) {
-  return (
-    <Table.Footer fullWidth>
-      <Table.Row>
-        <Table.HeaderCell></Table.HeaderCell>
-        {timeslots.map(timeslot => (
-          <FooterCell
-            participants={participants}
-            key={timeslot}
-            timeslot={timeslot}
-            interactive={interactive}
-            hovered={interactive && hoveredColumn === timeslot}
-            active={interactive && finalDate === timeslot}
-            onMouseEnter={interactive ? () => setHoveredColumn(timeslot) : null}
-            onMouseLeave={interactive ? () => setHoveredColumn(null) : null}
-          />
-        ))}
-      </Table.Row>
-    </Table.Footer>
-  );
-}
-
-TableFooter.propTypes = {
-  participants: PropTypes.arrayOf(
-    PropTypes.shape({
-      answers: PropTypes.objectOf(PropTypes.oneOf(['unavailable', 'available', 'ifneedbe'])),
-    })
-  ),
-  timeslots: PropTypes.array.isRequired,
-  interactive: PropTypes.bool.isRequired,
-  finalDate: PropTypes.string,
-  hoveredColumn: PropTypes.string,
-  setHoveredColumn: PropTypes.func,
-};
-
-export function AnswerGrid({
+export default function AnswerGrid({
   participant,
   user,
   name,
@@ -634,58 +279,4 @@ AnswerGrid.propTypes = {
   hasBusyTimes: PropTypes.bool,
   busyTimes: PropTypes.object.isRequired,
   duration: PropTypes.number.isRequired,
-};
-
-export function ParticipantGrid({finalDate, setFinalDate, isCreator, finalized}) {
-  const timeslots = useSelector(getNewdleTimeslots);
-  const participants = useSelector(getNewdleParticipants);
-  const [hoveredColumn, setHoveredColumn] = useState(null);
-
-  if (timeslots.length === 0) {
-    return null;
-  }
-
-  return (
-    <div className={styles['participant-grid']}>
-      <Table textAlign="center">
-        <TableHeader
-          timeslots={timeslots}
-          interactive={!finalized}
-          finalDate={finalDate}
-          setFinalDate={setFinalDate}
-          isCreator={isCreator}
-          hoveredColumn={hoveredColumn}
-          setHoveredColumn={setHoveredColumn}
-        />
-        <Table.Body>
-          {_.sortBy(participants, ['name']).map(participant => (
-            <SummaryRow
-              key={participant.id}
-              participant={participant}
-              timeslots={timeslots}
-              interactive={!finalized}
-              finalDate={finalDate}
-              hoveredColumn={hoveredColumn}
-              setHoveredColumn={setHoveredColumn}
-            />
-          ))}
-        </Table.Body>
-        <TableFooter
-          participants={participants}
-          timeslots={timeslots}
-          interactive={!finalized}
-          finalDate={finalDate}
-          hoveredColumn={hoveredColumn}
-          setHoveredColumn={setHoveredColumn}
-        />
-      </Table>
-    </div>
-  );
-}
-
-ParticipantGrid.propTypes = {
-  finalDate: PropTypes.string,
-  setFinalDate: PropTypes.func.isRequired,
-  isCreator: PropTypes.bool.isRequired,
-  finalized: PropTypes.bool.isRequired,
 };
